% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{build_app}
\alias{build_app}
\title{Build a GUI application of CircSeqAlignTk}
\usage{
build_app(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link[shiny]{shinyApp}}.}
}
\value{
A Shiny application object.
}
\description{
Build a graphical user interface (GUI) application for CircSeqAlignTk
using Shiny package.
}
\details{
The CircSeqAlignTk graphical user interface (GUI) application
is built using the Shiny package. Users need to install the Shiny package
and associated packages (shinyFiles, shinyjs) to use this application.
Additionally, the installation of RBowtie2 and Rhisat2 is required
to use the full functionality of this application.
}
\examples{
\dontrun{
library(shiny)
library(CircSeqAlignTk)

app <- build_app()
runApp(app)
}
}
