% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{package}
\name{CircSeqAlignTk-package}
\alias{CircSeqAlignTk}
\alias{CircSeqAlignTk-package}
\title{CircSeqAlignTk: End-to-End Analysis of Small RNA-Seq Data from Viroids}
\description{
CircSeqAlignTk is a toolkit for the analysis of RNA-Seq data derived from circular genome sequences, with a primary focus on viroids, circular RNAs typically consisting of a few hundred nucleotides. The toolkit supports an end-to-end analysis pipeline, from alignment to visualization.
}
\details{
Refer to the vignette for an overview of the package,
quick start, and detailed usages.
}
\examples{
browseVignettes("CircSeqAlignTk")

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bitdessin/CircSeqAlignTk}
  \item Report bugs at \url{https://github.com/bitdessin/CircSeqAlignTk/issues}
}

}
\author{
\strong{Maintainer}: Jianqiang Sun \email{sun@bitdessin.dev} (\href{https://orcid.org/0000-0002-3438-3199}{ORCID})

Other contributors:
\itemize{
  \item Xi Fu [contributor]
  \item Wei Cao [contributor]
}

}
