% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{CircSeqAlignTkCoverage-class}
\alias{CircSeqAlignTkCoverage-class}
\title{Class to save alignment coverage}
\description{
A class to store the alignment coverage
generated by \code{\link{calc_coverage}} function.
}
\section{Slots}{

\describe{
\item{\code{forward}}{A matrix containing the alignment coverage of
the forward strand reads.}

\item{\code{reversed}}{A matrix containing the alignment coverage of
the reversed strand reads.}

\item{\code{.figdata}}{A string of adapter sequence.}
}}

\seealso{
\code{\link{calc_coverage}}
}
