% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{CircSeqAlignTkRefIndex-class}
\alias{CircSeqAlignTkRefIndex-class}
\title{Class to store reference information}
\description{
A class to store reference information for alignment.
The object belongs to this class
is generated by \code{\link{build_index}} function.
}
\section{Slots}{

\describe{
\item{\code{name}}{Reference name.
The sequence name written the header of FASTA format file.}

\item{\code{seq}}{Reference sequence.}

\item{\code{length}}{Length of the reference sequence.}

\item{\code{fasta}}{A vector containing the paths to the two FASTA format files
of the type 1 and type 2 reference sequences, respectively.
See \code{\link{build_index}} for how FASTA format files are generated.}

\item{\code{index}}{A vector containing the paths to the two reference indexes
corresponding to the two FASTA format files
stored in the \code{fasta} slot, respectively.}

\item{\code{cut_loc}}{The position on the user-given sequence
(i.e., the type 1 sequence)
to cut for generating the type 2 reference sequence.}
}}

\seealso{
\code{\link{build_index}}
}
