% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{CircSeqAlignTkSim-class}
\alias{CircSeqAlignTkSim-class}
\title{Class to save information of synthetic reads}
\description{
A class to store parameters for generating synthetic sequence reads.
The object belongs to this class
is generated by \code{\link{generate_reads}} function.
}
\section{Slots}{

\describe{
\item{\code{seq}}{A string of a genome sequence,
which is used for sampling synthetic sequence reads.}

\item{\code{adapter}}{A string of adapter sequence.}

\item{\code{read_info}}{A data frame storing the summary information of read
generation. It contains the start and end positions of sampling, strand,
and nucleotide sequence of each synthetic read.}

\item{\code{peak}}{A data frame storing the peaks information of alignment coverage.}

\item{\code{coverage}}{A \code{\link{CircSeqAlignTkCoverage-class}}
storing the information of alignment coverage.}

\item{\code{fastq}}{A path to FASTQ format file saving the synthetic reads.}
}}

\seealso{
\code{\link{generate_reads}}, \code{\link{CircSeqAlignTkCoverage-class}}
}
