% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_slot_contents}
\alias{get_slot_contents}
\title{Get the slot contents from a formal class}
\usage{
get_slot_contents(object, name)
}
\arguments{
\item{object}{An object from a formally defined class.}

\item{name}{The name of the slot.}
}
\value{
The contents of the specified slot from the given object.
}
\description{
This function returns the slot contents from a formal class.
It is convenient to use \code{@} when accessing the contents of a slot,
however, using \code{@} will generate warnings
during the unit tests under software development.
This function was created to avoid that warning.
Users do not have to use this function.
}
\examples{
output_dpath <- tempdir()
sim <- generate_reads(output = file.path(output_dpath, 'sample1.fq.gz'))
head(get_slot_contents(sim, 'peak'))
}
