% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{package}
\name{CircSeqAlignTk-package}
\alias{CircSeqAlignTk}
\alias{CircSeqAlignTk-package}
\title{CircSeqAlignTk: A toolkit for end-to-end analysis of RNA-seq data for circular genomes}
\description{
CircSeqAlignTk is an end-to-end toolkit for RNA-Seq data analysis of circular genome sequences, supporting tasks from alignment to visualization. It is primarily designed for studying viroids, which are small, circular RNAs typically composed of a few hundred nucleotides. In addition to analysis capabilities, CircSeqAlignTk provides a streamlined interface for generating synthetic RNA-Seq data that closely mimics real datasets. This functionality enables developers to benchmark alignment tools, test novel alignment algorithms, and validate new workflows. This vignette offers an overview of CircSeqAlignTk and demonstrates its usage.
}
\details{
Refer to the vignette for an overview of the package,
quick start, and detailed usages.
}
\examples{
browseVignettes("CircSeqAlignTk")

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jsun/CircSeqAlignTk}
  \item Report bugs at \url{https://github.com/jsun/CircSeqAlignTk/issues}
}

}
\author{
\strong{Maintainer}: Jianqiang Sun \email{sun@bitdessin.dev} (\href{https://orcid.org/0000-0002-3438-3199}{ORCID})

Other contributors:
\itemize{
  \item Xi Fu [contributor]
  \item Wei Cao [contributor]
}

}
