% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{package}
\name{CircSeqAlignTk-package}
\alias{CircSeqAlignTk}
\alias{CircSeqAlignTk-package}
\title{CircSeqAlignTk: A toolkit for end-to-end analysis of RNA-seq data for circular genomes}
\description{
CircSeqAlignTk is an R package for end-to-end analysis, from alignment to visualization, of RNA-seq data obtained from organelles or organisms with circular genome sequences, such as bacteria, viruses, and viroids. In addition, CircSeqAlignTk implements a tidy interface to generate synthetic sequencing data that mimic real RNA-seq data, allowing developers to evaluate the performance of alignment tools, new alignment algorithms, and new workflows.
}
\details{
Refer to the vignette for an overview of the package,
quick start, and detailed usages.
}
\examples{
vignette("CircSeqAlignTk", package = "CircSeqAlignTk")

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jsun/vsrnaseqtk}
  \item Report bugs at \url{https://github.com/jsun/vsrnaseqtk/issues}
}

}
\author{
\strong{Maintainer}: Jianqiang Sun \email{sun@biunit.dev} (\href{https://orcid.org/0000-0002-3438-3199}{ORCID})

Authors:
\itemize{
  \item Xi Fu
  \item Wei Cao
}

}
