% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{package}
\name{CircSeqAlignTk-package}
\alias{CircSeqAlignTk}
\alias{CircSeqAlignTk-package}
\title{CircSeqAlignTk: A toolkit for end-to-end analysis of RNA-seq data for circular genomes}
\description{
CircSeqAlignTk is designed for end-to-end RNA-Seq data analysis of circular genome sequences, from alignment to visualization. It mainly targets viroids which are composed of 246-401 nt circular RNAs. In addition, CircSeqAlignTk implements a tidy interface to generate synthetic sequencing data that mimic real RNA-Seq data, allowing developers to evaluate the performance of alignment tools and workflows.
}
\details{
Refer to the vignette for an overview of the package,
quick start, and detailed usages.
}
\examples{
browseVignettes("CircSeqAlignTk")

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jsun/CircSeqAlignTk}
  \item Report bugs at \url{https://github.com/jsun/CircSeqAlignTk/issues}
}

}
\author{
\strong{Maintainer}: Jianqiang Sun \email{sun@biunit.dev} (\href{https://orcid.org/0000-0002-3438-3199}{ORCID})

Other contributors:
\itemize{
  \item Xi Fu [contributor]
  \item Wei Cao [contributor]
}

}
